Refactor code to utilize C# switch expressions
The existing code in the `SpravyCommandService.cs` file was refactored to switch from traditional switch-case statements to modern C# switch expressions. This change results in more readable and succint code. In addition, the CancellationToken was used directly instead of using `CancellationToken.None` and a new method `PostUi` was added in `ObjectExtension.cs`. These changes enhance the overall cleanliness and readability of the codebase.