Update code for entity update and styling
The changes mainly focus on updating the method of updating entity values in the code. The simpler method replaced the more complex one. Some modifications were also made in the Stepper.axaml, where changes were applied to the styles of the Grid and PathIcon. The database project also added a reference to the Microsoft.EntityFrameworkCore.Relational package.