Add SpravyFormat for string formatting and update dependencies
Implemented a new class, `SpravyFormat`, for string formatting with `IParameters`. Updated several models, classes, and files to replace the use of the "SmartFormat" package with the new `SpravyFormat` functionality. Consequently, removed the SmartFormat dependency from the project. This approach enhances code cohesion within the application, and removes reliance on an external package for this fundamental operation.