Add support for Android Arm64 runtime
A new Android Arm64 runtime support was added to the Runtime class in the Runtime.cs file, and the ProjectBuilderFactory.cs file was updated to include it. Minor changes were also made in the AndroidProjectBuilder.cs file to enable compilation, and in the Spravy.Ui.Android.csproj file, RuntimeIdentifiers was changed to RuntimeIdentifier for consistency.