Remove SukiUI, introduce new controls and update styles
This commit transitions from using the SukiUI library to using custom controls that are defined within the project. The such controls as BusyArea, GroupBox, and Stepper were introduced, and various files were updated to reflect these changes in dependencies and references. Multiple styles were also modified to be compatible with the new controls, and the Avalonia version was updated as well.