Update GroupBox header alignment in LoginView
The GroupBox in the LoginView has been updated with a CSS class for center alignment. This change includes an edit in the GroupBox.axaml file to introduce additional styles that allow for both left and center alignments of the header content. The header alignment can now be controlled by applying appropriate CSS classes.