Update PathControl to support separators
The PathControl has been updated to support customizable separators between segments, replacing the previous "Steps" concept with "Segments". Also, this refactoring included a change in all axaml files where the PathControl is used, replacing the "Steps" with "Segments". Now, the control is more flexible and capable of displaying a wider variety of paths.